requirements.txt
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cmdstanpy==1.1.0
contourpy==1.0.7
convertdate==2.4.0
cycler==0.11.0
ephem==4.1.4
exceptiongroup==1.1.0
filelock==3.9.0
fonttools==4.38.0
geographiclib==2.0
geopy==2.3.0
hijri-converter==2.2.4
holidays==0.19
huggingface-hub==0.12.0
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
matplotlib==3.6.3
newsapi-python==0.2.6
nltk==3.8.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pluggy==1.0.0
prophet==1.1.2
protobuf==4.21.12
pycodestyle==2.10.0
PyMeeus==0.5.12
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
scipy==1.10.0
six==1.16.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.0
tweepy==4.12.1
typing_extensions==4.4.0
urllib3==1.26.14
